+++ src/Form/ViewEditType.php
@@ -4,7 +4,7 @@ namespace App\Form;
 
 use App\Entity\Project;
 use App\Entity\View;
-use App\Service\ViewService;
+use App\Repository\WorklogRepository;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
@@ -13,7 +13,7 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class ViewEditType extends AbstractType
 {
-    public function __construct(protected readonly ViewService $viewService)
+    public function __construct(protected readonly WorklogRepository $worklogRepository)
     {
     }
 
@@ -24,15 +24,23 @@ class ViewEditType extends AbstractType
                 'class' => Project::class,
                 'choice_label' => 'name',
                 'multiple' => true,
-                'attr' => ['class' => 'form-element'],
+                'row_attr' => ['class' => 'form-row form-choices'],
+                'attr' => [
+                    'class' => 'form-element',
+                    'data-choices-target' => 'choices',
+                ],
                 'label' => 'view.projects',
                 'required' => false,
             ])
             ->add('workers', ChoiceType::class, [
-                'choices' => $this->viewService->getWorklogUsers(),
+                'choices' => $this->worklogRepository->getDistinctWorklogUsers(),
                 'mapped' => false,
                 'multiple' => true,
-                'attr' => ['class' => 'form-element'],
+                'row_attr' => ['class' => 'form-row form-choices'],
+                'attr' => [
+                    'class' => 'form-element',
+                    'data-choices-target' => 'choices',
+                ],
                 'label' => 'view.workers',
                 'required' => false,
             ]);