+++ src/Repository/WorklogRepository.php
@@ -6,8 +6,11 @@ use App\Entity\InvoiceEntry;
 use App\Entity\Project;
 use App\Entity\Worklog;
 use App\Model\Invoices\InvoiceEntryWorklogsFilterData;
+use App\Service\ViewService;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
+use Knp\Component\Pager\Pagination\PaginationInterface;
+use Knp\Component\Pager\PaginatorInterface;
 
 /**
  * @extends ServiceEntityRepository<Worklog>
@@ -19,7 +22,9 @@ use Doctrine\Persistence\ManagerRegistry;
  */
 class WorklogRepository extends ServiceEntityRepository
 {
-    public function __construct(ManagerRegistry $registry)
+    private const WORKLOGS_PAGINATOR_LIMIT = 50;
+
+    public function __construct(ManagerRegistry $registry, private readonly PaginatorInterface $paginator, private readonly ViewService $viewService)
     {
         parent::__construct($registry, Worklog::class);
     }
@@ -90,5 +95,72 @@ class WorklogRepository extends ServiceEntityRepository
         }
 
         return $qb->getQuery()->execute();
+    }
+
+    public function getTeamReportData(\DateTime $from, \DateTime $to, $viewId, $page): PaginationInterface
+    {
+        $view = $this->viewService->getViewFromId($viewId);
+        $parameters = [
+            'date_from' => $from->format('Y-m-d H:i:s'),
+            'date_to' => $to->format('Y-m-d H:i:s'),
+        ];
+
+        if (!empty($view)) {
+            $dataProviders = $view->getDataProviders()->getValues();
+            $projects = $view->getProjects()->getValues();
+            $workers = $view->getWorkers();
+        }
+
+        $qb = $this->createQueryBuilder('wor');
+        $qb->where($qb->expr()->between('wor.started', ':date_from', ':date_to'));
+        $qb->leftJoin('wor.project', 'project');
+        $qb->leftJoin('wor.issue', 'issue');
+        $qb->select('wor', 'project', 'issue');
+
+        if (!empty($projects)) {
+            $qb->andWhere('wor.project IN (:projects)');
+            $parameters['projects'] = [];
+            foreach ($projects as $project) {
+                $parameters['projects'][] = $project->getId();
+            }
+        }
+
+        if (!empty($dataProviders)) {
+            $qb->andWhere('wor.dataProvider IN (:dataProviders)');
+            $parameters['dataProviders'] = [];
+            foreach ($dataProviders as $dataProvider) {
+                $parameters['dataProviders'][] = $dataProvider->getId();
+            }
+        }
+
+        if (!empty($workers)) {
+            $qb->andWhere('wor.worker IN (:workers)');
+            $parameters['workers'] = $workers;
+        }
+
+        $qb->setParameters($parameters);
+
+        return $this->paginator->paginate(
+            $qb,
+            $page,
+            self::WORKLOGS_PAGINATOR_LIMIT,
+            ['defaultSortFieldName' => 'wor.started', 'defaultSortDirection' => 'asc']
+        );
+    }
+
+    public function getDistinctWorklogUsers(): array
+    {
+        $qb = $this->createQueryBuilder('wor');
+        $result = $qb
+            ->distinct()
+            ->select('wor.worker')
+            ->getQuery()->getResult();
+
+        $workers = [];
+        foreach ($result as $workLogWorker) {
+            $workers[$workLogWorker['worker']] = $workLogWorker['worker'];
+        }
+
+        return $workers;
     }
 }