+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed to using client->versionName to issue->version mapping for project billing.
 * Added project lead/mail to project.
 * Removed usused fields from account and client.
+* Add team report export
+* Add open spout extension
+* Add choices to views
 * Add team report
 * Add workers to views
 * Add view filtering to management reports